601-emulator/requirements.txt
azure-data-tables==12.*